+++ CHANGELOG.md
@@ -112,6 +112,9 @@ calculations.
 
  - #150: Further initializers for SuperNOVAS structures that can be used to empty initialize data.
 
+ - #154: Added `novas_print_hms()` and `novas_print_dms()` functions to convert time / angles to their HMS/DMS string 
+   representations, with the chosen separators, and the desired decimal places on the seconds.
+
  - #155: Added `novas_parse_iso_date()` for pasing ISO 8601 timestamps, including those expressed in the proleptic
    Gregorian calendar prior to the Gregorian calendar reform of 1582, for which the ISO 8601 timestamp differs from
    the astronomical calendar date assumed by `novas_parse_timestamp()`.